Update bb.nextflow.template
When you have the bb.nextflow.template create a new bucket it first does so in `S3Stack`. However, that `ExistingBucket=No` is propagated to `NextflowStack`. That stack again will try to create the S3 bucket and will fail because it was already created.

Changing line 335 to `ExistingBucket: Yes` fixes this. At this point, either the bucket was existing at launch or was already created as part of `S3Stack` so it is always existing by this point.